+++ 03-image-stats/cpp/ppm_stats.hpp
@@ -1,12 +1,13 @@
 #ifndef KV_PPM_STATS_HPP
 #define KV_PPM_STATS_HPP
 
+#include <cstdint>
 #include <expected>
 #include <iosfwd>
 #include <string>
 #include <string_view>
 
-#include "../../common/exit_codes.hpp"
+#include "../../common/cpp/exit_codes.hpp"
 
 enum class StatsError {
     kOk,
@@ -20,16 +21,16 @@ enum class StatsError {
 };
 
 struct Stats {
-    int width{};
-    int height{};
-    int max_val{};
-    long long pixel_count{};
-    long long total_r{};
-    long long total_g{};
-    long long total_b{};
+    int32_t width{};
+    int32_t height{};
+    int32_t max_val{};
+    int64_t pixel_count{};
+    int64_t total_r{};
+    int64_t total_g{};
+    int64_t total_b{};
     double y_min{};
     double y_max{};
-    int histogram[8]{};
+    int32_t histogram[8]{};
 };
 
 struct StatsResult {
